# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: 家居装修全流程管理平台 API
  version: 1.0.0
  description: 覆盖装修项目、设计阶段、材料清单、预算项与施工节点的全栈接口。
servers:
  - url: http://127.0.0.1:19305/api/v1
paths:
  /auth/login:
    post:
      summary: 用户登录
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          type: object
          properties:
            token: { type: string }
...